# doc/openapi.yaml
openapi: 3.0.0
info:
  title: 飲食店向け簡易オーダーシステム
  version: "1.0"
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /customer/dish:
    get:
      tags:
        - customer
      summary: メニューの取得
      operationId: customerMenuGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    category_name:
                      type: string
                      description: カテゴリ名
                    dishes:
                      type: array
                      items:
                        type: object
                        properties:
                          dish_id:
                            type: integer
                            description: 料理ID
                          dish_name:
                            type: string
                            description: 料理名
                          image_url:
                            type: string
                            description: 料理画像のURL
                          dish_price:
                            type: integer
                            description: 料理の価格
                          available_num:
                            type: integer
                            description: 料理の在庫数
        "401":
          description: Unauthorized (トークンが無効)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: qr_code_token
          description: 座席のQRコードに含まれているトークン
  /customer/dish/{dish_id}:
    get:
      tags:
        - customer
      summary: 料理の詳細を取得
      operationId: customerMenuDishGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dish_id:
                    type: integer
                    description: 料理ID
                  dish_name:
                    type: string
                    description: 料理名
                  image_url:
                    type: string
                    description: 料理画像のURL
                  dish_price:
                    type: integer
                    description: 料理の価格
                  available_num:
                    type: integer
                    description: 料理の在庫数
        "400":
          description: Bad Request (dish_idに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: qr_code_token
          description: 座席のQRコードに含まれているトークン
        - schema:
            type: integer
          in: path
          required: true
          name: dish_id
          description: 料理のID
  /customer/order:
    get:
      tags:
        - customer
      summary: 注文履歴、合計金額の取得
      operationId: customerOrderGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: integer
                    description: 注文ID
                  ordered_dishes:
                    type: array
                    items:
                      type: object
                      properties:
                        ordered_dish_id:
                          type: integer
                          description: 注文した料理の注文ID
                        ordered_dish_name:
                          type: string
                          description: 注文した料理の名称
                        ordered_dish_price:
                          type: integer
                          description: 注文した料理の価格
                        ordered_dish_num:
                          type: integer
                          description: 注文した料理の個数
                        ordered_dish_status:
                          type: string
                          description: 注文した料理のステータス
                  total_price:
                    type: integer
                    description: 合計金額
                  is_paid:
                    type: boolean
                    description: 会計済みかどうか
                  created_at:
                    type: string
                    description: 注文日時
        "401":
          description: Unauthorized (トークンが無効)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: qr_code_token
          description: 座席のQRコードに含まれているトークン

    post:
      tags:
        - customer
      summary: 注文・追加注文送信
      operationId: customerOrderPost
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: integer
                    description: 注文ID
                  ordered_dishes:
                    type: array
                    items:
                      type: object
                      properties:
                        ordered_dish_id:
                          type: integer
                          description: 注文した料理の注文ID
                        ordered_dish_name:
                          type: string
                          description: 注文した料理の名称
        "400":
          description: Bad Request (注文リクエストに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ordered_dishes:
                  type: array
                  items:
                    type: object
                    properties:
                      dish_id:
                        type: integer
                        description: 注文した料理のID
                      quantity:
                        type: integer
                        description: 注文した料理の数量
                    required:
                      - dish_id
                      - quantity
        required: true
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: qr_code_token
          description: 座席のQRコードに含まれているトークン
  /customer/order/finish:
    post:
      tags:
        - customer
      summary: 注文確定・会計依頼
      description: お会計ボタンを押すとそこで注文が締め切られる。is_order_finishedの状態を変更する(false → true)。
      operationId: customerOrderFinish
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request (注文が既に締め切られている、もしくは未注文)
        "401":
          description: Unauthorized (トークンが無効)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: qr_code_token
          description: 座席のQRコードに含まれているトークン
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT